web/server/requirements.txt
astroid==1.6.4
certifi==2018.4.16
colorama==0.3.9
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.1
pytz==2018.5
six==1.11.0
Django==2.1.1
wheel==0.31.1
djongo==1.2.29
requests==2.19.1
djangorestframework==3.8.2
python-dotenv==0.9.1
django-cors-headers
pymongo
halo
2to3
future
simplejson
python-dateutil
dnspython
requests-futures
